requirements.txt
Django==2.0.8
djangorestframework==3.9.0
django-taggit==0.22.2
